Test.Cardano.Prelude

Documentation

runTests :: [IO Bool] -> IO ()

discoverPropArg :: TExpQ (a -> Group)

discoverRoundTrip :: TExpQ Group

discoverRoundTripArg :: TExpQ (a -> Group)

roundTripsAesonShow :: (Eq a, MonadTest m, ToJSON a, FromJSON a, Show a) => a -> m ()

roundTripsAesonBuildable :: (Eq a, MonadTest m, ToJSON a, FromJSON a, Buildable a) => a -> m ()

Round trip any a with both ToJSON and FromJSON instances.

roundTripsCanonicalJsonPretty :: (Eq a, Show a, MonadTest m, ToJSON Identity a, FromJSON (Either SchemaError) a) => a -> m ()

Pretty round trip any a with both ToJSON and FromJSON canonical instances.

trippingBuildable :: forall f a b m. (HasCallStack, Buildable (f a), Eq (f a), Show b, Applicative f, MonadTest m) => a -> (a -> b) -> (b -> f a) -> m ()

Round trip using given encode and decode functions for types with a Buildable instance

Various properties and predicates

qcIsJust :: Maybe a -> Property

qcIsNothing :: Show a => Maybe a -> Property

qcIsLeft :: Show b => Either a b -> Property

qcIsRight :: Show a => Either a b -> Property

qcElem :: (Show a, Eq a, Show (t a), Foldable t) => a -> t a -> Property

qcNotElem :: (Show a, Eq a, Show (t a), Foldable t) => a -> t a -> Property

qcFail :: Text -> Property

A property that is always false

Monadic properties

assertProperty :: Monad m => Bool -> Text -> PropertyM m ()

Call stopProperty if boolean value is false.

stopProperty :: Monad m => Text -> PropertyM m a

Stop PropertyM execution with given reason. The property will fail.

maybeStopProperty :: Monad m => Text -> Maybe a -> PropertyM m a

Use stopProperty if the value is Nothing or return something it the value is Just.

splitIntoChunks :: Monad m => Word -> [a] -> PropertyM m [NonEmpty a]

Split given list into chunks with size up to given value. TODO: consider using `sumEquals maxSize (length items)`

expectedOne :: Monad m => Text -> [a] -> PropertyM m a

HSpec utils

expectationError :: Text -> Expectation

Generators

splitWord :: Word64 -> Word64 -> Gen [Word64]

Split given integer total into parts parts TODO: improve naming!

sumEquals :: Word64 -> Word64 -> Gen [Word64]

Generate list of arbitrary positive integers which sum equals given sum. All elements in the list will be smaller or equal then first parameter

Helpers

(.=.) :: (Eq b, Show b) => (a -> b) -> (a -> b) -> a -> Property infixr 5

Extensional equality combinator. Useful to express function properties as functional equations.

(>=.) :: (Show (m b), Eq (m b)) => (a -> m b) -> (a -> m b) -> a -> Property infixr 5

Monadic extensional equality combinator.

shouldThrowException :: Exception e => (a -> b) -> Selector e -> a -> Expectation

Semigroup/monoid laws

formsSemigroup :: (Show m, Eq m, Semigroup m) => m -> m -> m -> Property

formsMonoid :: (Show m, Eq m, Semigroup m, Monoid m) => m -> m -> m -> Property

formsCommutativeMonoid :: (Show m, Eq m, Semigroup m, Monoid m) => m -> m -> m -> Property

class Nonrepeating a where

Type for generating list of unique (nonrepeating) elemets.

Methods

nonrepeating :: Int -> Gen [a]

class ArbitraryUnsafe a where

ArbitraryUnsafe class ~~~~~~~~~~~~~~~~~~~~~~~~

This class is the same as Arbitrary, except instances of this class for stuff like public/secret keys, VSS shares, commitments etc. are designed not to mimic real data as presisely as possible (using OpenSSL random), but rather to be simple and efficient.

This is especially useful for benchmarking.

Note: we don't need Generic to derive instances of ArbitraryUnsafe. We can either use one-line instance declaration, or -XStandaloneDeriving or -XDeriveAnyClass to write something like deriving (Arbitrary, ArbitraryUnsafe).

Minimal complete definition

Nothing

Methods

arbitraryUnsafe :: Gen a

default arbitraryUnsafe :: Arbitrary a => Gen a

newtype SmallGenerator a

Constructors

SmallGenerator

Fields

getSmallGenerator :: a

makeSmall :: Gen a -> Gen a

sublistN :: Int -> [a] -> Gen [a]

Choose a random (shuffled) subset of length n. Throws an error if there's not enough elements.

arbitrarySizedS :: Int -> Gen ByteString

Make arbitrary ByteString of given length.

arbitrarySizedSL :: Int -> Gen ByteString

Make arbitrary Lazy of given length.

runGen :: Gen a -> a

Get something out of a quickcheck Gen without having to do IO

assertIsLeftConstr :: (Buildable b, Data a, HasCallStack, MonadTest m) => Constr -> Either a b -> m ()

assertIsRight :: (Buildable a, HasCallStack, MonadTest m) => Either a b -> m ()

assertIsJust :: (HasCallStack, MonadTest m) => Maybe a -> m ()

assertIsNothing :: (Buildable a, HasCallStack, MonadTest m) => Maybe a -> m ()

compareValueRight :: (Buildable a, Eq b, HasCallStack, MonadTest m, Show b) => b -> Either a b -> m ()

discoverGolden :: TExpQ Group

eachOf :: (Show a, HasCallStack) => TestLimit -> Gen a -> (a -> PropertyT IO ()) -> Property

Check that eachOf testLimit generated things hasProperty

goldenTestCanonicalJSONDec :: (Eq a, FromJSON (Either SchemaError) a, HasCallStack, Show a) => a -> FilePath -> Property

Check that Canonical JSON decodes to the datatype

goldenTestJSONDec :: (Eq a, FromJSON a, HasCallStack, Show a) => a -> FilePath -> Property

Only check that the datatype equals the decoding of the file

goldenTestJSON :: (Eq a, FromJSON a, HasCallStack, Show a, ToJSON a) => a -> FilePath -> Property

goldenTestJSONPretty :: (Eq a, FromJSON a, HasCallStack, Show a, ToJSON a) => a -> FilePath -> Property

getText :: Int -> Int -> Text

genBytes :: Int -> Gen ByteString

genUTF8Byte :: Gen ByteString

gen32Bytes :: Gen ByteString

genWord32 :: Gen Word32

genWord16 :: Gen Word16

genNatural :: Gen Natural

genNominalDiffTime :: Gen NominalDiffTime

decodeBase16 :: ByteString -> Maybe ByteString

Decode a given ByteString which was originally encoded using encode or encodeWithIndex.

encodeBase16 :: ByteString -> ByteString

Encodes a given ByteString to base-16 and line wraps every 16 bytes.

encodeWithIndex :: ByteString -> ByteString

Encodes a given ByteString to base-16 and displays it alongside its byte offset (line wrapped every 16 bytes).