Test.Cardano.Prelude.Golden

Description

Helper functions for use in golden testing of datatypes

Documentation

discoverGolden ∷ TExpQ Group

eachOf ∷ (Show a, HasCallStack) ⇒ TestLimit → Gen a → (a → PropertyT IO ()) → Property

Check that eachOf testLimit generated things hasProperty

goldenTestCanonicalJSONDec ∷ (Eq a, FromJSON (Either SchemaError) a, HasCallStack, Show a) ⇒ a → FilePath → Property

Check that Canonical JSON decodes to the datatype

goldenTestJSONDec ∷ (Eq a, FromJSON a, HasCallStack, Show a) ⇒ a → FilePath → Property

Only check that the datatype equals the decoding of the file

goldenTestJSON ∷ (Eq a, FromJSON a, HasCallStack, Show a, ToJSON a) ⇒ a → FilePath → Property

goldenTestJSONPretty ∷ (Eq a, FromJSON a, HasCallStack, Show a, ToJSON a) ⇒ a → FilePath → Property

getText ∷ Int → Int → Text