Test.Cardano.Prelude.QuickCheck.Arbitrary

Description

A number of helpers for QuickCheck, used in the old codebase

TODO: Decide whether or not to remove this module

Documentation

class Nonrepeating a where

Type for generating list of unique (nonrepeating) elemets.

Methods

nonrepeating ∷ Int → Gen [a]

class ArbitraryUnsafe a where

ArbitraryUnsafe class ~~~~~~~~~~~~~~~~~~~~~~~~

This class is the same as Arbitrary, except instances of this class for stuff like public/secret keys, VSS shares, commitments etc. are designed not to mimic real data as presisely as possible (using OpenSSL random), but rather to be simple and efficient.

This is especially useful for benchmarking.

Note: we don't need Generic to derive instances of ArbitraryUnsafe. We can either use one-line instance declaration, or -XStandaloneDeriving or -XDeriveAnyClass to write something like deriving (Arbitrary, ArbitraryUnsafe).

Minimal complete definition

Nothing

Methods

arbitraryUnsafe ∷ Gen a

arbitraryUnsafe ∷ Arbitrary a ⇒ Gen a

newtype SmallGenerator a

Constructors

SmallGenerator

Fields

getSmallGenerator ∷ a

makeSmall ∷ Gen a → Gen a

sublistN ∷ Int → [a] → Gen [a]

Choose a random (shuffled) subset of length n. Throws an error if there's not enough elements.

arbitrarySizedS ∷ Int → Gen ByteString

Make arbitrary ByteString of given length.

arbitrarySizedSL ∷ Int → Gen ByteString

Make arbitrary Lazy of given length.

runGen ∷ Gen a → a

Get something out of a quickcheck Gen without having to do IO