Test.Cardano.Prelude.QuickCheck.Property

Description

This module contains a number of helpers for writing QuickCheck properties

TODO: Decide whether or not to remove this from the codebase

Various properties and predicates

qcIsJust ∷ Maybe a → Property

qcIsNothing ∷ Show a ⇒ Maybe a → Property

qcIsLeft ∷ Show b ⇒ Either a b → Property

qcIsRight ∷ Show a ⇒ Either a b → Property

qcElem ∷ (Show a, Eq a, Show (t a), Foldable t) ⇒ a → t a → Property

qcNotElem ∷ (Show a, Eq a, Show (t a), Foldable t) ⇒ a → t a → Property

qcFail ∷ Text → Property

A property that is always false

Monadic properties

assertProperty ∷ Monad m ⇒ Bool → Text → PropertyM m ()

Call stopProperty if boolean value is false.

stopProperty ∷ Monad m ⇒ Text → PropertyM m a

Stop PropertyM execution with given reason. The property will fail.

maybeStopProperty ∷ Monad m ⇒ Text → Maybe a → PropertyM m a

Use stopProperty if the value is Nothing or return something it the value is Just.

splitIntoChunks ∷ Monad m ⇒ Word → [a] → PropertyM m [NonEmpty a]

Split given list into chunks with size up to given value. TODO: consider using `sumEquals maxSize (length items)`

expectedOne ∷ Monad m ⇒ Text → [a] → PropertyM m a

HSpec utils

expectationError ∷ Text → Expectation

Generators

splitWord ∷ Word64 → Word64 → Gen [Word64]

Split given integer total into parts parts TODO: improve naming!

sumEquals ∷ Word64 → Word64 → Gen [Word64]

Generate list of arbitrary positive integers which sum equals given sum. All elements in the list will be smaller or equal then first parameter

Helpers

(.=.) ∷ (Eq b, Show b) ⇒ (a → b) → (a → b) → a → Property infixr 5

Extensional equality combinator. Useful to express function properties as functional equations.

(>=.) ∷ (Show (m b), Eq (m b)) ⇒ (a → m b) → (a → m b) → a → Property infixr 5

Monadic extensional equality combinator.

shouldThrowException ∷ Exception e ⇒ (a → b) → Selector e → a → Expectation

Semigroup/monoid laws

formsSemigroup ∷ (Show m, Eq m, Semigroup m) ⇒ m → m → m → Property

formsMonoid ∷ (Show m, Eq m, Semigroup m, Monoid m) ⇒ m → m → m → Property

formsCommutativeMonoid ∷ (Show m, Eq m, Semigroup m, Monoid m) ⇒ m → m → m → Property