Test.Cardano.Prelude.Base16

Description

Functions to encode and decode ByteStrings in base 16 with line wrapping

This encoding is used for producing pretty output from golden tests. The golden values are stored on disk as indexed, line-wrapped hex dumps. These can then be nicely diffed to provide some indication of where serialisation errors may have occurred.

Documentation

decodeBase16 ∷ ByteString → Maybe ByteString

Decode a given ByteString which was originally encoded using encode or encodeWithIndex.

encodeBase16 ∷ ByteString → ByteString

Encodes a given ByteString to base-16 and line wraps every 16 bytes.

encodeWithIndex ∷ ByteString → ByteString

Encodes a given ByteString to base-16 and displays it alongside its byte offset (line wrapped every 16 bytes).